pipeline/requirements.txt
snowflake-connector-python>=3.7.0
python-dotenv>=1.0.0
anthropic>=0.40.0
